backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.2
tweepy==4.14.0
google-auth==2.25.2
google-api-python-client==2.110.0
facebook-sdk==3.1.0
requests==2.31.0
python-dotenv==1.0.0
alembic==1.13.0
bcrypt==4.1.2
